Changed the type from inputlen from uint8_t to uint32_t  since data can be larger then 255-bytes. (Bill Cox found this bug)

+++ src/catena-blake2b.c
@@ -3,7 +3,7 @@
 #include "hash.h"
 
 
-inline void __Hash1(const uint8_t *input, const uint8_t inputlen,
+inline void __Hash1(const uint8_t *input, const uint32_t inputlen,
 		      uint8_t hash[H_LEN])
 {
   blake2b_state ctx;
